Create Pareto Front
Finds the Pareto front given either randomly generated data for the values of the objective functions or specific data given by the user. By default, assumes minimisation is desired for both objectives, but this may be changed easily. Outputs the Pareto matrix and the volume of the (first) Pareto shell.